# chore(perf): Improve performance