--- a/lib/formatters.js
+++ b/lib/formatters.js
@@ -20,9 +20,9 @@
 
 var util = require('util'),
     colors = require('colors/safe'),
-    stringify = require('json-stringify-safe'),
+    safeStringify = require('safe-json-stringify'),
     serializeErr = require('serr'),
-    omit = require('lodash').omit;
+    _ = require('lodash');
 
 var DEFAULT_NOT_AVAILABLE = 'n/a';
 
@@ -97,7 +97,7 @@ function formatDevTrace(level, context, message, args, err) {
     str += '\n' + colorize(colors.gray, serializeErr(err).toString(printStack));
   }
 
-  var localContext = omit(context, formatDevTrace.omit);
+  var localContext = _.omit(context, formatDevTrace.omit);
   str += Object.keys(localContext).length ?
       ' ' + colorize(colors.gray, util.inspect(localContext)) :
       '';
@@ -188,14 +188,24 @@ function formatJsonTrace(level, context, message, args, err) {
   return formatJsonTrace.stringify(formatJsonTrace.toObject(level, context, message, args, err));
 }
 
-formatJsonTrace.stringify = stringify;
+formatJsonTrace.stringify = (obj) => {
+  try {
+    return JSON.stringify(obj);
+  } catch (err) {
+    return safeStringify(obj);
+  }
+};
 formatJsonTrace.toObject = (level, context, message, args, err) => {
-  const printStack = API.stacktracesWith.indexOf(level) > -1;
+  let log = {};
 
-  return Object.assign({}, context, {
-    time: new Date().toISOString(),
-    lvl: level,
-    err: err && serializeErr(err).toObject(printStack),
-    msg: util.format(message, ...args)
-  });
+  for (let attrname in context) {
+    log[attrname] = context[attrname];
+  }
+
+  log.time = new Date();
+  log.lvl = level;
+  log.err = err && serializeErr(err).toObject(API.stacktracesWith.indexOf(level) > -1);
+  log.msg = util.format.apply(global, [message].concat(args));
+
+  return log;
 };