--- a/lib/logops.js
+++ b/lib/logops.js
@@ -24,7 +24,7 @@ module.exports = logops();
 function noop() {};
 
 function logops(options) {
-  let opts = Object.assign({
+  let opts = merge({
     level: process.env.LOGOPS_LEVEL || DEFAULT_LEVEL,
     format: formatters[process.env.LOGOPS_FORMAT] || (process.env.NODE_ENV === 'development' ?
       formatters.dev :
@@ -43,39 +43,39 @@ function logops(options) {
    *   ['DEBUG', 'INFO', 'WARN', 'ERROR', 'FATAL']
    */
   function logWrap(level) {
-    let context, message, args, trace, err;
+    return function log() {
+      let context, message, args, trace, err;
 
-    if (arguments[1] instanceof Error) {
-      // log.<level>(err, ...)
-      context = API.getContext();
-      args = Array.prototype.slice.call(arguments, 2);
-      if (!args.length) {
-        // log.<level>(err)
-        err = arguments[1];
-        message = err.name + ': ' + err.message;
+      if (arguments[0] instanceof Error) {
+        // log.<level>(err, ...)
+        context = API.getContext();
+        args = Array.prototype.slice.call(arguments, 1);
+        if (!args.length) {
+          // log.<level>(err)
+          err = arguments[0];
+          message = err.name + ': ' + err.message;
+        } else {
+          // log.<level>(err, "More %s", "things")
+          // Use the err as context information
+          err = arguments[0];
+          message = arguments[1];
+          args = Array.prototype.slice.call(args, 1);
+        }
+      } else if (arguments[0] == null || (typeof (arguments[0]) !== 'object' && arguments[0] !== null) || Array.isArray(arguments[0])) {
+        // log.<level>(msg, ...)
+        context = API.getContext();
+        message = arguments[0];
+        args = Array.prototype.slice.call(arguments, 1);
       } else {
-        // log.<level>(err, "More %s", "things")
-        // Use the err as context information
-        err = arguments[1];
-        message = arguments[2];
-        args = Array.prototype.slice.call(args, 1);
+        // log.<level>(fields, msg, ...)
+        context = merge(API.getContext(), arguments[0]);
+        message = arguments[1];
+        args = Array.prototype.slice.call(arguments, 2);
       }
-    } else if ((typeof (arguments[1]) !== 'object' && arguments[1] !== null) ||
-        arguments[1] == null ||
-        Array.isArray(arguments[1])) {
-      // log.<level>(msg, ...)
-      context = API.getContext();
-      message = arguments[1];
-      args = Array.prototype.slice.call(arguments, 2);
-    } else {
-      // log.<level>(fields, msg, ...)
-      context = Object.assign({}, API.getContext(), arguments[1]);
-      message = arguments[2];
-      args = Array.prototype.slice.call(arguments, 3);
-    }
 
-    trace = API.format(level, context || {}, message, args, err);
-    API.stream.write(trace + '\n');
+      trace = API.format(level, context || {}, message, args, err);
+      API.stream.write(trace + '\n');
+    }
   }
 
   /**
@@ -92,22 +92,12 @@ function logops(options) {
     levels.forEach((logLevel) => {
       let fn;
       if (logLevelIndex <= levels.indexOf(logLevel)) {
-        fn = logWrap.bind(global, logLevel);
+        fn = logWrap(logLevel);
       } else {
         fn = noop;
       }
       API[logLevel.toLowerCase()] = fn;
     });
-  }
-
-  /**
-   * Gets the current log level.
-   *
-   * @return {String}   One of the following values
-   *     ['DEBUG', 'INFO', 'WARN', 'ERROR', 'FATAL']
-   */
-  function getLevel() {
-    return opts.level;
   }
 
   /**
@@ -147,7 +137,7 @@ function logops(options) {
      * @return {String}   One of the following values
      *     ['DEBUG', 'INFO', 'WARN', 'ERROR', 'FATAL']
      */
-    getLevel: getLevel,
+    getLevel: () => opts.level,
 
     /**
      * Gets the context for a determinate trace. By default, this is a noop that
@@ -247,7 +237,7 @@ function logops(options) {
      * // {"hostname":"host.local","app":"server","time":"2015-12-23T11:47:25.862Z","lvl":"INFO","msg":"Startup"}
      * ```
      */
-    child: function(localContext) {
+    child: (localContext) => {
       // the current limitation for this approach is that child
       // loggers will overwrite the parent and localcontext when
       // calling againg `child.setContextGetter(fn)` or assign a function
@@ -256,7 +246,7 @@ function logops(options) {
       // considerer it a corner case and dont optimize for the use case
       return logops({
         level: API.getLevel(),
-        getContext: () => Object.assign({}, API.getContext(), localContext),
+        getContext: () => merge(API.getContext(), localContext),
         format: API.format,
         stream: API.stream,
       });
@@ -265,4 +255,24 @@ function logops(options) {
 
   setLevel(opts.level);
   return API;
+}
+
+/**
+ * Merges accesible properties in two objects.
+ * obj2 takes precedence when common properties are found
+ *
+ * @param {Object} obj1
+ * @param {Object} obj2
+ * @returns {{}} The merged, new, object
+ */
+function merge(obj1, obj2) {
+  var res = {}, attrname;
+
+  for (attrname in obj1) {
+    res[attrname] = obj1[attrname];
+  }
+  for (attrname in obj2) {
+    res[attrname] = obj2[attrname];
+  }
+  return res;
 }