--- a/lib/logops.js
+++ b/lib/logops.js
@@ -61,7 +61,8 @@ function logops(options) {
           message = arguments[1];
           args = Array.prototype.slice.call(args, 1);
         }
-      } else if (arguments[0] == null || (typeof (arguments[0]) !== 'object' && arguments[0] !== null) || Array.isArray(arguments[0])) {
+      } else if (arguments[0] == null || (typeof (arguments[0]) !== 'object' && arguments[0] !== null) ||
+          Array.isArray(arguments[0])) {
         // log.<level>(msg, ...)
         context = API.getContext();
         message = arguments[0];
@@ -75,7 +76,7 @@ function logops(options) {
 
       trace = API.format(level, context || {}, message, args, err);
       API.stream.write(trace + '\n');
-    }
+    };
   }
 
   /**
@@ -133,9 +134,6 @@ function logops(options) {
 
     /**
      * Gets the current log level.
-     *
-     * @return {String}   One of the following values
-     *     ['DEBUG', 'INFO', 'WARN', 'ERROR', 'FATAL']
      */
     getLevel: () => opts.level,