Several performance issues where found caused by using mostly `Object.assign` 

With Logging level enabled (writes logs)

| Type |  Before | After |
| ----- | ------- | ----- |
| `tee -a out.log > /dev/null` | 24,378 ops/sec ±4.59% | 39,560 ops/sec ±3.00%  | 
| `> out.log`  |  22,625 ops/sec ±2.96% | 43,136 ops/sec ±1.31% |
| `> /dev/null` | 26,333 ops/sec ±2.31% | 52,947 ops/sec ±1.80% |


With Logging level disabled (does not writes logs)

| Type |  Before | After |
| ----- | ------- | ----- |
| `tee -a out.log > /dev/null` | 67,893,971 ops/sec ±3.33% | 73,150,310 ops/sec ±1.64%  | 
| `> out.log`  |  69,832,990 ops/sec ±2.16% | 80,439,813 ops/sec ±1.17%  |
| `> /dev/null` | 69,598,851 ops/sec ±2.08% | 77,479,942 ops/sec ±1.47% |